chore: refactor, move exceptionHandling features from GrpcStatus to DefaultGrpcExceptionHandlerFunction
